+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.8.0 - TBD
 
++ Added `Restore-OpenADObject` to restore a deleted Active Directory object, from the Recycle Bin or as a tombstone when the Recycle Bin is not enabled
+
 ## v0.7.0 - 2026-08-27
 
 + Added `Add-OpenADGroupMember` and `Remove-OpenADGroupMember` to manage AD group members